Renaming main -> render
